+++ tests/SharpLouis.Tests/TableSearchPathTests.cs
@@ -2,7 +2,7 @@ using System;
 using System.IO;
 using System.Text;
 using AccessMind.SharpLouis;
-using FluentAssertions;
+using AwesomeAssertions;
 using Xunit;
 
 namespace AccessMind.SharpLouis.Tests;